WIP: Attempt dataloader implementation
Relates to #1370. A proof-of-concept dataloader implementation based on the
nestjs-dataloader library. In benchmarks it has no appreciable benefit based on
the "product list" benchmark and using dataloader to batch the productVariant
calls. Might have benefits in other scenarios but not sure currently whether it is
worth the added complexity.

In light of 063b5fe, the n+1 problem gets solved in many cases by joining the
nested data in the top-level query by inspecting the graphql fields requested.